+++ src/main/java/com/vieiracode/springbootcrud/repository/BooksRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.vieiracode.springbootcrud.model.BooksModel;
 
-public interface BooksRepository extends CrudRepository<BooksModel, Integer> {
+public interface BooksRepository extends CrudRepository<BooksModel, Long> {
 
 }
